requirements.txt
django==5.2
djangorestframework==3.16
django-extensions==4.1
django-allauth[socialaccount]==65.7
factory-boy==3.3
markdown==3.8
pytest-django==4.11
